% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_predict.R
\name{create_predict}
\alias{create_predict}
\title{Create data to predict IUCN status on}
\usage{
create_predict(data_full)
}
\arguments{
\item{data_full}{The output from the missForest_applied function}
}
\value{
A dataframe with all species to predict IUCN status on
}
\description{
This functions selects all the species for which the IUCN info is not available
}
