% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_species.R
\name{create_species}
\alias{create_species}
\title{Create data with only species information}
\usage{
create_species(data_full)
}
\arguments{
\item{data_full}{The output from the missForest_applied function}
}
\value{
A dataframe with species inforamtion
}
\description{
This functions creates data with species information
}
