% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_train_C.R
\name{create_train_C}
\alias{create_train_C}
\title{Create train data with all classified species}
\usage{
create_train_C(data_full)
}
\arguments{
\item{data_full}{The output from the missForest_applied function}
}
\value{
A dataframe with all classified species
}
\description{
This functions selects all the classified species for training
}
