% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_train_NC.R
\name{create_train_NC}
\alias{create_train_NC}
\title{Create train data with all non classified species}
\usage{
create_train_NC(data_full)
}
\arguments{
\item{data_full}{The output from the missForest_applied function}
}
\value{
A dataframe with all non classified species
}
\description{
This functions selects all the non classified species for training
}
