% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Split data into downsampled data for model}
\usage{
data_prep(data_NC, data_C)
}
\arguments{
\item{data_NC}{Your filled out data with no NA with all non classified species}

\item{data_C}{Your filled out data with no NA with all classified species}
}
\value{
A list with each downsampled dataframe from your data
}
\description{
This function downsamples your data to have even number of classified and non classified species into even bits based on your data
This dataset will be used to train the model
}
