% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCN_consensus.R
\name{IUCN_consensus}
\alias{IUCN_consensus}
\title{Consensus of IUCN status based on the 200 models}
\usage{
IUCN_consensus(data_predicted, splits, baseline)
}
\arguments{
\item{data_predicted}{The output from the IUCN_predict function}

\item{splits}{Length of the data_prep list}

\item{baseline}{A percentage, between 50 and 100\%, to keep IUCN prediction}
}
\value{
A data frame with its IUCN classifications, the number of models that classified it there and the percentage
}
\description{
This functions calculates the consensus of the 200 models based on a specified baseline
}
