% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCN_divide.R
\name{IUCN_divide}
\alias{IUCN_divide}
\title{Try out the missForest performance}
\usage{
IUCN_divide(data)
}
\arguments{
\item{data}{Your data with species as rownames, traits as column names and the IUCN column.
In this column, Species with no info are labelled NA and the others according to their status: CR,LC,NT...}
}
\value{
Your dataframe with IUCN column transformed into two categories : NT and T. Thr corresponds to species
in CR, EN or VU category and NThr species in LC or NT category
}
\description{
This functions tests the missForest on complete data using cross validation
}
