% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCN_test.R
\name{IUCN_test}
\alias{IUCN_test}
\title{Test the model on complete data}
\usage{
IUCN_test(data_split)
}
\arguments{
\item{data_split}{The output from the data_prep function}
}
\value{
A data frame with variable importance of each trait by permutation

A data frame with model performance : accuracy, Kappa, TSS

A data frame with Balanced accuracy across categoris
}
\description{
This functions tests the model on your data based on complete data using cross validation
}
