% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missForest.R
\name{missForest_applied}
\alias{missForest_applied}
\title{Run missForest based on its performance}
\usage{
missForest_applied(data_tofill, baseline, mf_test)
}
\arguments{
\item{data_tofill}{Your data with the NA with species as rownames, traits as columns}

\item{baseline}{Between 0.5 and 1, baseline of how well the missForest performed to predict}

\item{mf_test}{The output from the missForest_test function}
}
\value{
Your data with NAs filled out for traits where the missForest performed well
}
\description{
This functions runs the missForest on your data based on its performance for each trait
}
