% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missForest_test.R
\name{missForest_test}
\alias{missForest_test}
\title{Try out the missForest performance}
\usage{
missForest_test(data_noNA)
}
\arguments{
\item{data_noNA}{Data without NA with species as rownames, traits as columns}
}
\value{
A dataframe with R² or Accuracy from missForest cv according to the trait type
}
\description{
This functions tests the missForest on complete data using cross validation
}
