% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missForest_test.R
\name{missForest_test}
\alias{missForest_test}
\title{Try out the missForest performance}
\usage{
missForest_test(data)
}
\arguments{
\item{data}{The output from the IUCN predict function}
}
\value{
A dataframe with R² or Accuracy from missForest cv according to the trait type
}
\description{
This functions tests the missForest on complete data using cross validation
}
